migraet int64 and uint64 from long long and unsigned long long to int64_t and uint64_t

+++ CLKernel.h
@@ -13,8 +13,8 @@
 
 typedef int int32;
 typedef unsigned int uint32;
-typedef long long int64;
-typedef unsigned long long uint64;
+typedef int64_t int64;
+typedef uint64_t uint64;
 
 class EasyCL_EXPORT CLKernel {
     EasyCL *easycl; // NOT owned by this object, dont delete!
